chore: update to match formatting conventions
