+++ src/main/java/exceptions/TaskNotFoundException.java
@@ -1,4 +1,4 @@
-package Exceptions;
+package exceptions;
 
 public class TaskNotFoundException extends Exception {
   public TaskNotFoundException() {
