+++ src/main/java/utilities/DateTime.java
@@ -1,4 +1,4 @@
-package Utilities;
+package utilities;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
