+++ src/main/java/utilities/IO.java
@@ -1,4 +1,4 @@
-package Utilities;
+package utilities;
 
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
